feat: delete employee by id and AddDepartment

+++ human-resource-management/Controller/EmployeeController.cs
@@ -88,8 +88,22 @@ namespace human_resource_management.Controller
 
         public void DeleteEmployee(EmployeeModel employee)
         {
-            employeeRepository.Delete(employee);
-            Console.WriteLine("Employee deleted successfully.");
+            public void DeleteEmployee()
+        {
+            Console.Write("Nhập ID nhân viên cần xóa: ");
+            int id = int.Parse(Console.ReadLine() ?? string.Empty);
+            EmployeeModel employee = employeeRepository.GetById(id);
+            if  (id == null)
+            {
+                Console.WriteLine("Không tìm thấy nhân viên có ID: " + id);
+            }
+            else
+            {
+                employeeRepository.Delete(employee);
+                Console.WriteLine("Xóa nhân viên thành công.");
+            }
+
+        }
         }
 
         public void SortEmployeesBy(Func<EmployeeModel, IComparable> keySelector)
